fix(cli): split resolution probe counters
